requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.3
django-cors-headers==4.6.0
django-prometheus==2.3.1
djangorestframework==3.15.2
flake8==7.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
prometheus_client==0.21.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.2
